libflash/blocklevel: add blocklevel flags.
It has become apparent that communication between blocklevel and its
backend may be necessarily, at least so that the backend can inform
blocklevel as to if an erase must be done before a write.

An erase before flag isn't strictly necessary as erases can be performed
regardless of whether they need to be or not, however, the caveat with that
is that when erases don't need to be performed, this is likely due to the
backend not having erase blocks and therefore it may be impossible to set a
sane value for erase blocks which would be used to erase before write.

This flag saves backends that don't need erases before write to have to lie
about an erase block size.

Reviewed-By: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>