LEDS: Rename led node as leds
Presently we are creating '/ibm,opal/led' node and populating all LED
related information there. Ben [1] suggested to rename this as 'leds'.

Present code is included in skiboot skiboot-5.0 release..But we don't
have any consumer yet. Hence I think its fine to make this changes.

Finally updated node name in doc file.

[1] https://lists.ozlabs.org/pipermail/linuxppc-dev/2015-June/130433.html

Suggested-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Reviewed-by: Stewart Smith <stewart@linux.vnet.ibm.com>
[stewart@linux.vnet.ibm.com: We're *lucky* that no FSP based system shipped
with skiboot with /led/ rather than /leds/, in future, as the OpenPower ecosystem
grows, we will unlikly be able to make this kind of assumption that nobody
else went and used this layout. This is likely the last of this kind of change.]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>