external/boot-tests: Fix logging functions in boot_test.sh
Fix the error() and msg() functions in boot_test.sh so they don't cause a
shell error when called before $target is set (i.e. before parameter
parsing).

Change error() to allow the use of multiple arguments in the message, in
line with msg(). Use $* rather than $@ as it's very slightly more correct.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>