IPMI: Only log events that needs attention
Presently we are logging all the events to service processor (FSP/BMC).
But on BMC machines we should only log events that requires attention.

As per PEL spec, we should log events with severity >= 0x22 and "service
action flag" is "on". But in our case, all logs OPAL originagted logs
are makred as report externally.

So lets log all events that are originated from OAPL (presently all logs
as payload is not logging any PEL event) and  severity >= 0x22.

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Cc: Ananth N Mavinakayanahalli <ananth@in.ibm.com>
Cc: Alistair Popple <alistair@popple.id.au>
Cc: Jeremy Kerr <jk@ozlabs.org>
[stewart@linux.vnet.ibm.com: s/SEL:/IPMI:/ in prlog]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>