pci: Don't create -T location codes on non-FSP machines
This is a temporary workaround. We disable the code that adds the
"-T" portion of the location code to network devices on non-FSP machines.

This code is a bit fragile, and it does incorrect things on dual-slot
mezzanines such as the one in Habanero. If we need to support this
on OpenPower, we will need to add a concept of sibling slots to the
infrastructure to properly handle the numbering.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>