external/gard: Check the validity of the flash
If the GUARD partition has been corrupted such that ECC checks fail and not
even the first gard record can be read then the gard tool will be unable to
do anything. If this happens it is possible that hostboot will have a
similar problem.

The only sane thing to do at is point is probably to wipe the GUARD
partition.

This patch adds a check for this case and provides the user with the option
to wipe the entirety of the GUARD partition to attempt recovery.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>