external/common: Write an includeable rules.mk.
It would be nice if tools like the gard tool or pflash don't have to worry
about how to build the arch specific code they want to include through the
new external/common code.

This patch adds an external/common/rules.mk which each tool can include and
with some minor tweaking of the existing makefiles it should get the arch
code building nicely.

The one caveat is that it requires a symlink in the directory to create
common/ dir for everything to behave.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>