phb3: Handle fence in phb3_pci_msi_check_q to fix hang
If the PHB is fenced during phb3_pci_msi_check_q, it can get stuck in an
infinite loop waiting to lock the FFI. Further, as the phb lock is held
during this function it will prevent any other CPUs from dealing with
the fence, leading to the entire system hanging.

If the PHB_FFI_LOCK returns all Fs, return immediately to allow the
fence to be dealt with.

Signed-off-by: Ian Munsie <imunsie@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>