external/gard: Update record clearing to match Hostboots expectations
When the gard tool is told to clear all the gard records it actually erases
the entire flash and inserts ECC bytes only for the size of the first gard
record.

The current method appears to have stopped working, Hostboot expects the
entire partition to have ECC bytes throughout the entire partition, and as
such the current method causes Hostboot to error and be unable to bring up
machines.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Reviewed-by: Samuel Mendoza-Jonas <sam.mj@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>