hw/bt.c: Timeout messages when bt interface isn't functional
During system bring up we may not have a properly functioning ipmi
interface. This prevents skiboot completing the boot process as it
waits for certain bt messages to complete before continuing.

This patch alters the bt message timeouts to ensure messages timeout
in the case of a non-responsive bt interface.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>