hw/fake-rtc: Add support for emulated rtc clock
This patch adds support for an emulated rtc clock over existing fake rtc
implementation for generic platform (e.g BML). Presently a fake rtc
clock is initialized when reserved region named 'ibm,fake-rtc' is
present in the boot device tree. This mem-region points to the initial
value of bcd coded date-time values. However as this region is in system
memory hence its not updated with time. This results in an error from
hwclock tool which tries to detect a change in system clock and then
complains "Timed out waiting for time change."

The patch overcomes this issue by emulating an rtc clock whose date-time
values are calculated from the difference of current timebase and its
value when the initial epoch was assigned. The initial epoch is set from
the values at "ibm,fake-rtc" memory region.

Acked-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Vaibhav Jain <vaibhav@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>