console: Completely flush output buffer before power down and reboot
Completely flush the output buffer of the console driver before
power down and reboot.  Implements the flushing function for uart
consoles, which includes the astbmc and rhesus platforms.

Adds a new function, flush(), to the con_ops struct that allows
each console driver to specify how their output buffers are flushed.

In the cec_power_down and cec_reboot functions, the flush function
of the driver is called if it exists.

This fixes an issue where some console output is sometimes lost before
power down or reboot in uart consoles. If this issue is also prevalent
in other console types then it can be fixed later by adding a .flush
to that driver's con_ops.

Signed-off-by: Russell Currey <ruscur@russell.cc>
[stewart@linux.vnet.ibm.com: reduce diff size, change flush function name]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>