opal-prd: display explicit message on IBM Power systems
Today, when run on an IBM Power systems, opal-prd complains in syslog
with a set of messages similar to these :

	opal-prd: CTRL: Starting PRD daemon
	opal-prd: I2C: Found Chip: 00000000 engine 1 port 0
	opal-prd: I2C: Found Chip: 00000010 engine 1 port 0
	opal-prd: CTRL: Listening on control socket /run/opal-prd-control
	opal-prd: FW: Can't open PRD device /dev/opal-prd: No such file or directory
	opal-prd: FW: Error initialising PRD channel
	opal-prd: CTRL: stopping PRD daemon

Which are difficult to interpret for a person not initiated to Power
firmware.

The patch below detects if the platform has support for PRD by looking
at the device tree property :

	/sys/firmware/devicetree/base/ibm,opal/diagnostics/compatible

and stops opal-prd early in the main routine with an explicit
message for the user.

Signed-off-by: Cédric Le Goater <clg@fr.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>