Use VALGRIND callout rather than tricky malloc/memcpy
The tricky memcpy trips up various static analysers and looks rather
odd. The more correct thing to do is to ask valgrind to mark it
as undefined. There's a call for that, and we depend on valgrind
to run the test suite anyway, so we're not introducing a new
dependency.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>