hw/bt: Remove state machine and use only B2H_ATN and H2B_ATN.
Correct use of B2H_ATN and H2B_ATN means that we don't need to track
states.

The only caveat is that currently we use states to know if we need
to send a message to the BMC, change this to simply seeing if the top
message has been sent or not.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Acked-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>