pflash: Remove ast.h
pflash contained a copy of the include/ast.h header. It had grown stale,
so remove in and link in the common header.

Note that it's hard to test that we haven't broken tools in the
external/ directory these days; when making changes we need to test with
amd64, ppc64, ppc64le and arm to ensure that everything can build!

Signed-off-by: Joel Stanley <joel@jms.id.au>
Reviewed-by: Cédric Le Goater <clg@fr.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>