Fix up extract-gcov for gcc > 4.8
The story of extract-gcov is not necessarily a pleasant one, involving
GCC internals, padding of data structures, differences in data structures
that are designed to change whenever GCC wants to and a strong desire
to not implement a VFS in skiboot or some other streaming interface
(and associated userspace and other such blergh).

This patch makes us be all explicit about padding in the structures,
enabling -Wpadding for extract-gcov.c.

We also get all strict over the size of things and add support for
gcc 5.1, which added an extra counter.

There is likely GCC hacking in my future to make this a lot less
fragile.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>