Add OPAL_CONSOLE_FLUSH to the OPAL API
uart consoles only flush output when polled.  The Linux kernel calls
these pollers frequently, except when in a panic state.  As such, panic
messages are not fully printed unless the system is configured to reboot
after panic.

This patch adds a new call to the OPAL API to flush the buffer.  If the
system has a uart console (i.e. BMC machines), it will incrementally
flush the buffer, returning if there is more to be flushed or not.  If
the system has a different console, the function will have no effect.
This will allow the Linux kernel to ensure that panic message have been
fully printed out.

The existing synchronous flushing mechanism used in OPAL's shutdown and
reboot routines has been refactored into a helper that repeatedly calls
the new partial flush function.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>