libflash/blocklevel: Add keep_alive parameter
Currently the file backend will keep a file descriptor open until the
structure is destroyed. This is undesirable for daemons and long running
processes that only have a very occasional need to access the file. Keeping
the file open requires users of blocklevel to close and reinit their
structures every time, doing is isn't disastrous but can easily be
managed at the interface level.

This has been done at the blocklevel_device interface so as to provide
minimal changes to arch_flash_init().

At the moment there isn't a need for the actually flash driver to be able
to do this, this remains unimplmented there.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>